requirements/base.txt
Django==4.2.5
django-environ==0.11.2
frontend==0.6
mysql-connector-python==8.1.0
djangorestframework==3.14.0
django-mptt==0.15.0
Pillow==10.0.1
drf-yasg==1.21.7
colorlog==6.7.0
django-cleanup==8.0.0
